NER_Image_Classification/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0
datasets==2.13.1
nltk==3.8.1
scikit-learn==1.2.2
Pillow==9.5.0
tensorflow==2.12.0
tf-keras==2.12.0
numpy==1.24.3